# examples/java-spring-complex/openapi.yaml
openapi: 3.0.3
info:
  title: E-Commerce Orders API
  description: |
    Comprehensive RESTful API for an e-commerce orders platform.
    Supports user management, product catalog, order lifecycle, payments,
    refunds, shipments, reviews, and administrative operations.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.orders.example.com
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      tags: [Products]
      operationId: listProducts
      summary: List products
      description: Returns a paginated, filterable list of active products. Public endpoint.
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          description: Full-text search on name and description
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
            format: double
        - name: maxPrice
          in: query
          schema:
            type: number
            format: double
        - name: sortBy
          in: query
          schema:
            type: string
            default: name
            enum: [name, price, createdAt]
        - name: sortDir
          in: query
          schema:
            type: string
            default: asc
            enum: [asc, desc]
      responses:
        '200':
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'

    post:
      tags: [Products]
      operationId: createProduct
      summary: Create product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: SKU already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{id}:
    get:
      tags: [Products]
      operationId: getProductById
      summary: Get product by ID
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Products]
      operationId: updateProduct
      summary: Update product (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Products]
      operationId: deleteProduct
      summary: Delete product (admin only)
      description: Soft-deletes the product by marking it inactive.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /products/{id}/reviews:
    get:
      tags: [Reviews, Products]
      operationId: getProductReviews
      summary: Get reviews for a product
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of reviews for the product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewPage'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────
  # Orders
  # ─────────────────────────────────────────
...
  /payments:
    post:
      tags: [Payments]
      operationId: initiatePayment
      summary: Initiate payment for an order
      description: Creates a payment authorization for an order. Payment starts in AUTHORIZED state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiatePaymentRequest'
      responses:
        '201':
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Order already paid or cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payments/{id}:
    get:
      tags: [Payments]
      operationId: getPaymentById
      summary: Get payment details
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/{id}/capture:
    post:
      tags: [Payments]
      operationId: capturePayment
      summary: Capture an authorized payment
      description: |
        Captures a previously authorized payment and moves the order to PROCESSING.
        Business Rule: payment-required-before-shipment (captured payments enable shipment)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Payment captured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Payment is not in AUTHORIZED state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payments/{id}/void:
    post:
      tags: [Payments]
      operationId: voidPayment
      summary: Void an authorized payment
      description: Voids a PENDING or AUTHORIZED payment. Cannot void captured payments (use refund instead).
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Payment voided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Payment cannot be voided in current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─────────────────────────────────────────
  # Refunds
  # ─────────────────────────────────────────
...
  /shipments/{id}:
    get:
      tags: [Shipments]
      operationId: getShipmentById
      summary: Get shipment details
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Shipment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /shipments/{id}/status:
    put:
      tags: [Shipments]
      operationId: updateShipmentStatus
      summary: Update shipment status (admin only)
      description: |
        Updates the shipment status. Moving to SHIPPED requires a captured payment.
        Business Rule: payment-required-before-shipment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShipmentStatusRequest'
      responses:
        '200':
          description: Shipment status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Payment not captured or invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─────────────────────────────────────────
  # Reviews
  # ─────────────────────────────────────────
  /reviews:
    post:
      tags: [Reviews]
      operationId: createReview
      summary: Submit a product review
      description: |
        Allows authenticated users to review a product.
        Business Rule: review-requires-completed-order — user must have a completed order containing the product.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User has no completed order for this product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─────────────────────────────────────────
  # Admin
  # ─────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from the authentication endpoint
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Resource unique identifier
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        stockQuantity:
          type: integer
        category:
          type: string
        sku:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateProductRequest:
      type: object
      required: [name, price, stockQuantity, category, sku]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
          minimum: 0.01
        stockQuantity:
          type: integer
          minimum: 0
        category:
          type: string
        sku:
          type: string
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        stockQuantity:
          type: integer
        category:
          type: string
    ProductPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    # ── Order ──
...
    Payment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        status:
          type: string
          enum: [PENDING, AUTHORIZED, CAPTURED, VOIDED, REFUNDED, FAILED]
        amount:
          type: number
          format: double
        paymentMethod:
          type: string
        transactionId:
          type: string
        gatewayReference:
          type: string
        capturedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    InitiatePaymentRequest:
      type: object
      required: [orderId, paymentMethod]
      properties:
        orderId:
          type: integer
          format: int64
        paymentMethod:
          type: string
          enum: [CREDIT_CARD, DEBIT_CARD, PAYPAL, BANK_TRANSFER, CRYPTO]
        cardToken:
          type: string
          description: Tokenized card reference from payment gateway

    # ── Refund ──
...
    Shipment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        status:
          type: string
          enum: [PENDING, READY_TO_SHIP, SHIPPED, IN_TRANSIT, OUT_FOR_DELIVERY, DELIVERED, FAILED, RETURNED]
        carrier:
          type: string
        trackingNumber:
          type: string
        shippingMethod:
          type: string
        estimatedDelivery:
          type: string
          format: date-time
        shippedAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
    UpdateShipmentStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [PENDING, READY_TO_SHIP, SHIPPED, IN_TRANSIT, OUT_FOR_DELIVERY, DELIVERED, FAILED, RETURNED]
        trackingNumber:
          type: string
        carrier:
          type: string
        estimatedDelivery:
          type: string
          format: date-time

    # ── Review ──
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        author:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateReviewRequest:
      type: object
      required: [productId, orderId, rating]
      properties:
        productId:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          maxLength: 2000
    ReviewPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        totalElements:
          type: integer
        totalPages:
          type: integer

    # ── Admin ──
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string